requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
scikit-learn>=1.0.0
pandas>=2.0.0
pyarrow>=14.0.0  # for parquet support
certifi>=2024.2.2
beautifulsoup4>=4.12.0
requests>=2.31.0
docarray>=0.21.0 
lxml>=5.0.0